+++ src/libs/github-page-detect.js
@@ -1,7 +1,8 @@
 // Functions referenced and modified from Sindre Sorhus's `refined-github` repository
 // Source: https://github.com/sindresorhus/refined-github/blob/master/source/libs/page-detect.ts
 
-import { getCleanPathname, getRepoPath } from './utils';
+import { getCleanPathname } from './utils';
+import { getRepoPath } from './github-utils';
 
 export const isGitHub = () => location.hostname === 'github.com';
